[C-2923] Fix toasts in modal screens
