+++ packages/mobile/src/components/core/ModalScreen.tsx
@@ -1,0 +1,17 @@
+import type { ReactNode } from 'react'
+
+import { Toasts } from '../toasts'
+
+type ModalScreenProps = {
+  children: ReactNode
+}
+
+export const ModalScreen = (props: ModalScreenProps) => {
+  const { children } = props
+  return (
+    <>
+      <Toasts />
+      {children}
+    </>
+  )
+}
