+++ packages/mobile/src/components/core/index.ts
@@ -39,3 +39,4 @@ export * from './KeyboardAvoidingView'
 export * from './InputErrorMessage'
 export * from './DogEar'
 export * from './LockedStatusBadge'
+export * from './ModalScreen'
