+++ packages/mobile/src/screens/edit-playlist-screen/index.ts
@@ -1,3 +1,3 @@
-export * from './EditPlaylistScreen'
+export * from './EditPlaylistModalScreen'
 export * from './LegacyEditPlaylistScreen'
 export * from './CreatePlaylistScreen'
