+++ packages/mobile/src/screens/tip-artist-screen/index.ts
@@ -1,1 +1,1 @@
-export * from './TipArtistModal'
+export * from './TipArtistModalScreen'
