+++ packages/mobile/src/screens/upload-screen/screens/index.ts
@@ -1,4 +1,4 @@
-export * from './UploadScreen'
+export * from './UploadModalScreen'
 export * from './CompleteTrackScreen'
 export * from './UploadingTracksScreen'
 export * from './UploadCompleteScreen'
