+++ packages/mobile/src/screens/wallet-connect/WalletConnectModalScreen.tsx
@@ -1,6 +1,6 @@
 import { createNativeStackNavigator } from '@react-navigation/native-stack'
 
-import { Toasts } from 'app/components/toasts'
+import { ModalScreen } from 'app/components/core'
 
 import { useAppScreenOptions } from '../app-screen/useAppScreenOptions'
 
@@ -11,17 +11,16 @@ const Stack = createNativeStackNavigator()
 
 const screenOptionOverrides = { headerRight: () => null }
 
-export const WalletConnectStack = () => {
+export const WalletConnectModalScreen = () => {
   const screenOptions = useAppScreenOptions(screenOptionOverrides)
 
   return (
-    <>
-      <Toasts />
+    <ModalScreen>
       <Stack.Navigator screenOptions={screenOptions}>
         <Stack.Screen name='Wallets' component={WalletConnectScreen} />
       </Stack.Navigator>
       <WalletsDrawer />
       <ConfirmRemoveWalletDrawer />
-    </>
+    </ModalScreen>
   )
 }
